Soluciona #7: mejora la estructura del repositorio siguiendo las mejores prácticas del lenguaje
