+++ monumento/ubicacion.go
@@ -1,0 +1,6 @@
+package monumento
+
+type Ubicacion struct {
+    Calle  string
+    Numero string
+}
